Move renderCandidatePartyList to utils.tsx for better separation
