+++ libs/ui/src/ui_strings/index.ts
@@ -1,3 +1,4 @@
 export * from './app_strings';
 export * from './election_strings';
 export * from './ui_strings_context';
+export * from './utils';
